requirements.txt
loguru>=0.5.3
apscheduler>=3.6.3
peewee-extra-fields>=2.8.2
bit-help @ https://github.com/daveusa31/bit-help/archive/54b93f94946ff1f4c122aeb0e7d02dcc1559bd3f.zip
wrapcache @ https://github.com/daveusa31/wrapcache/archive/be06f720b0fbadd737dd547d22bb4faa803ee4a5.zip
yandex-money-sdk @ https://github.com/daveusa31/yandex-money-sdk-python/archive/94b40b926d7ecf9f7bf153d32b82bd5ae4e2699c.zip
